// client/src/generated/client/models/RsOpenWizardRound.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Play with friends
 * Play with friends
 *
 * The version of the OpenAPI document: 1.0
 * Contact: theo.dm94@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    RsImmutableRoundState,
    RsImmutableRoundStateFromJSON,
    RsImmutableRoundStateFromJSONTyped,
    RsImmutableRoundStateToJSON,
} from './RsImmutableRoundState';
import {
    RsRoundStateForPlayer,
    RsRoundStateForPlayerFromJSON,
    RsRoundStateForPlayerFromJSONTyped,
    RsRoundStateForPlayerToJSON,
} from './RsRoundStateForPlayer';

/**
 * 
 * @export
 * @interface RsOpenWizardRound
 */
export interface RsOpenWizardRound {
    /**
     * 
     * @type {RsImmutableRoundState}
     * @memberof RsOpenWizardRound
     */
    immutableRoundState: RsImmutableRoundState;
    /**
     * 
     * @type {RsRoundStateForPlayer}
     * @memberof RsOpenWizardRound
     */
    currentRoundStateForPlayer: RsRoundStateForPlayer;
}

export function RsOpenWizardRoundFromJSON(json: any): RsOpenWizardRound {
    return RsOpenWizardRoundFromJSONTyped(json, false);
}

export function RsOpenWizardRoundFromJSONTyped(json: any, ignoreDiscriminator: boolean): RsOpenWizardRound {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'immutableRoundState': RsImmutableRoundStateFromJSON(json['immutableRoundState']),
        'currentRoundStateForPlayer': RsRoundStateForPlayerFromJSON(json['currentRoundStateForPlayer']),
    };
}

export function RsOpenWizardRoundToJSON(value?: RsOpenWizardRound | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'immutableRoundState': RsImmutableRoundStateToJSON(value.immutableRoundState),
        'currentRoundStateForPlayer': RsRoundStateForPlayerToJSON(value.currentRoundStateForPlayer),
    };
}