// client/src/generated/client/models/RsWizardGameStateForPlayer.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Play with friends
 * Play with friends
 *
 * The version of the OpenAPI document: 1.0
 * Contact: theo.dm94@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    RsFinishedWizardRound,
    RsFinishedWizardRoundFromJSON,
    RsFinishedWizardRoundFromJSONTyped,
    RsFinishedWizardRoundToJSON,
} from './RsFinishedWizardRound';
import {
    RsOpenWizardRound,
    RsOpenWizardRoundFromJSON,
    RsOpenWizardRoundFromJSONTyped,
    RsOpenWizardRoundToJSON,
} from './RsOpenWizardRound';

/**
 * 
 * @export
 * @interface RsWizardGameStateForPlayer
 */
export interface RsWizardGameStateForPlayer {
    /**
     * 
     * @type {Array<RsFinishedWizardRound>}
     * @memberof RsWizardGameStateForPlayer
     */
    oldRounds: Array<RsFinishedWizardRound>;
    /**
     * 
     * @type {RsOpenWizardRound}
     * @memberof RsWizardGameStateForPlayer
     */
    openWizardRound?: RsOpenWizardRound;
}

export function RsWizardGameStateForPlayerFromJSON(json: any): RsWizardGameStateForPlayer {
    return RsWizardGameStateForPlayerFromJSONTyped(json, false);
}

export function RsWizardGameStateForPlayerFromJSONTyped(json: any, ignoreDiscriminator: boolean): RsWizardGameStateForPlayer {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'oldRounds': ((json['oldRounds'] as Array<any>).map(RsFinishedWizardRoundFromJSON)),
        'openWizardRound': !exists(json, 'openWizardRound') ? undefined : RsOpenWizardRoundFromJSON(json['openWizardRound']),
    };
}

export function RsWizardGameStateForPlayerToJSON(value?: RsWizardGameStateForPlayer | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'oldRounds': ((value.oldRounds as Array<any>).map(RsFinishedWizardRoundToJSON)),
        'openWizardRound': RsOpenWizardRoundToJSON(value.openWizardRound),
    };
}